+++ motrix_envs/src/motrix_envs/locomotion/go1/walk_stairs_terrain.py
@@ -287,7 +287,7 @@ class Go1WalkStairsTask(NpEnv):
             high=self.cfg.commands.vel_limit[1],
             size=(num_envs, 3),
         )
-        return commands
+        return commands.astype(np.float32)
 
     def update_reward(self, state: NpEnvState) -> NpEnvState:
         data = state.data
